+++ js/xdata.js
@@ -1,0 +1,11 @@
+
+class Information {
+  constructor(Id, url, description, likes, comments){
+    this.Id = Id;
+    this.url = url;
+    this.description = description;
+    this.likes = likes;
+    this.comments = comments;
+  }
+}
+export default Information;
